+++ src/services/networkWatcher/mutation.ts
@@ -1,0 +1,5 @@
+export default `mutation($input: [AddazureNetworkWatcherInput!]!) {
+  addazureNetworkWatcher(input: $input, upsert: true) {
+    numUids
+  }
+}`;